requirements.txt
openai==0.28.0
pandas
numpy
streamlit
python-dotenv 
docx2txt
google.generativeai
PyPDF2
nltk  
requests
spacy
matplotlib
seaborn
scikit-learn
plotly
diffusers
transformers
torch
torchvision
accelerate
streamlit
safetensors
Pillow
